+++ src/components/Summary/index.tsx
@@ -7,7 +7,9 @@ import { useContext } from "react";
 
 
 export function Summary(){
-    const data = useContext(TransactionsContext);
+    const transactions = useContext(TransactionsContext);
+
+    console.log(transactions)
 
     return (             
         <Container>
